+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Improve create group admin api [#452](https://github.com/rokwire/groups-building-block/issues/452)
+
 ## [1.39.0] - 2024-04-30
 ### Added
 - Improve group events APIs [#450](https://github.com/rokwire/groups-building-block/issues/450)